+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.4.0
 - Add support for glob pattern in variable file paths ([#4](https://github.com/qetza/replacetokens/issues/4)).
+- Add support for YAML variable files ([#5](https://github.com/qetza/replacetokens/issues/5)).
 - Add `parseVariables` function to parse string variables (load, flatten & merge).
 - Removed `flattenAndMerge` function (use `parseVariables` instead).